+++ framework/views/carousel.js
@@ -240,6 +240,35 @@ limitations under the License.
       },
 
       /**
+       * @param {Number} ratio
+       */
+      setAutoScrollRatio: function(ratio) {
+        if (ratio < 0.0 || ratio > 1.0) {
+          throw new Error('Invalid ratio.');
+        }
+
+        this._element[0].setAttribute('auto-scroll-ratio', ratio);
+      },
+
+      /**
+       * @return {Number}
+       */
+      getAutoScrollRatio: function(ratio) {
+        var attr = this._element[0].getAttribute('auto-scroll-ratio');
+
+        if (!attr) {
+          return 0.5;
+        }
+
+        var ratio = parseFloat(attr);
+        if (ratio < 0.0 || ratio > 1.0) {
+          throw new Error('Invalid ratio.');
+        }
+
+        return isNaN(ratio) ? 0.5 : ratio;
+      },
+
+      /**
        * @param {Number} index
        * @param {Object} [options]
        * @param {Function} [options.callback]
@@ -452,7 +481,7 @@ limitations under the License.
           var duration = 0.3;
           var scrollDelta = duration * 100 * velocity;
           var scroll = this._scroll + (this._getScrollDelta(this._lastDragEvent) > 0 ? -scrollDelta : scrollDelta);
-          scroll = this._normalizeScrollPosition(scroll);
+          scroll = this._normalizeScrollPosition(scroll)
 
           this._scroll = scroll;
 
@@ -491,6 +520,15 @@ limitations under the License.
 
             return left - right;
           });
+
+          var lastScroll = this._lastActiveIndex * size,
+            scrollRatio = Math.abs(scroll - lastScroll) / size;
+
+          if (arr[0] === lastScroll && scrollRatio >= this.getAutoScrollRatio()) {
+            if (arr.length > 1) {
+              return arr[1];
+            }
+          }
 
           return arr[0];
         } else {